Added a new util module
This will contain the MAC functions for now.
Other functions may be liberally added.
