+++ keysign/GetKeySection.py
@@ -35,7 +35,6 @@ from monkeysign.gpg import GpgRuntimeError
 from compat import gtkbutton
 import Keyserver
 from SignPages import ScanFingerprintPage, SignKeyPage, PostSignPage
-from SignPages import mac_verify
 
 import key
 
@@ -46,7 +45,8 @@ from gi.repository import GdkX11
 from gi.repository import GstVideo
 
 from compat import monkeysign_expired_keys, monkeysign_revoked_keys
-import key
+
+from .util import mac_verify
 
 
 Gst.init([])
